# shared/openapi.yaml
openapi: 3.1.0
info:
  title: Zhihu Hackathon Demo API
  version: 0.2.0
  description: >
    Shared contract for the current monorepo scaffold. GET /api/search remains
    the implemented search mapping endpoint and must not be broken. The P0
    product-layer contract for the AI persona experience is POST
    /api/demo/search; persona chat is planned for P0.5 at POST
    /api/personas/chat.
servers:
  - url: http://localhost:8000
    description: Local backend
...
paths:
...
  /api/search:
    get:
      tags:
        - Search
      summary: Frontend-friendly search endpoint
      description: >
        Implemented. This is the existing frontend-friendly Zhihu search
        mapping endpoint and remains available for compatibility, debugging
        and product-layer recall. The P0 AI persona product contract is POST
        /api/demo/search. Without a real Zhihu key, this endpoint returns a
        clear JSON error instead of crashing.
      operationId: getSearch
      x-implementation-status: implemented
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 1
          example: 不工作了能去哪儿
        - name: count
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5
          example: 1
      responses:
        "200":
          description: Search succeeded and was mapped into frontend-friendly fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchSuccessResponse"
              examples:
                success:
                  value:
                    success: true
                    data:
                      query: 不工作了能去哪儿
                      count: 1
                      hasMore: false
                      searchHashId: ""
                      items:
                        - id: zhihu_item_1
                          type: answer
                          title: 裸辞以后可以去哪里
                          text: 一段来自知乎内容或 mock 内容的摘要。
                          url: https://www.zhihu.com/question/mock/answer/mock
                          author:
                            name: 知乎作者
                            avatar: ""
                            badge: ""
                            badgeText: ""
                          stats:
                            commentCount: 0
                            voteUpCount: 0
                            rankingScore: 0
                          comments: []
                          editTime: 0
                          authorityLevel: ""
                          source:
                            provider: zhihu
                            url: https://www.zhihu.com/question/mock/answer/mock
                          evidence:
                            text: 一段来自知乎内容或 mock 内容的摘要。
                            source:
                              provider: zhihu
                              url: https://www.zhihu.com/question/mock/answer/mock
        "400":
          description: Missing or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "500":
          description: Missing local Zhihu credentials or other backend error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
              examples:
                missingKey:
                  value:
                    success: false
                    error:
                      code: ZHIHU_AUTH_FAILED
                      message: 知乎 API 鉴权失败
...
  /api/agent/search:
    post:
      tags:
        - Agent
      summary: Start a lightweight demo search agent task
      description: >
        Creates an in-memory task, starts the existing demo search pipeline in
        the background, and returns immediately with a taskId. The final result
        keeps the existing demo.v1 response shape used by /api/demo/search.
      operationId: startDemoSearchAgent
      x-implementation-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemoSearchRequest"
      responses:
        "200":
          description: Agent task accepted and running.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentTaskStartResponse"
        "400":
          description: Missing or invalid request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/agent/tasks:
    get:
      tags:
        - Agent
      summary: List recent persistent Agent tasks
      description: >
        Read-only debug helper for the internal Agent Runtime debug page.
        Returns recent Postgres-backed Agent tasks ordered by createdAt
        descending. Does not include artifacts or events; use
        /api/agent/tasks/{taskId} for the full snapshot.
      operationId: listPersistentAgentTasks
      x-implementation-status: implemented-debug
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Recent persistent Agent tasks.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersistentAgentTaskListResponse"
        "503":
          description: DATABASE_URL is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    post:
      tags:
        - Agent
      summary: Create a persistent Agent task
      description: >
        Persistent Agent Runtime. Creates a task in Postgres, writes
        task.created, enqueues a BullMQ job, marks it queued, and writes
        task.enqueued. A separate worker consumes the job and runs the
        rule-based business stage workflow: understand_goal_rule,
        plan_search_llm, retrieve_sources, normalize_candidates, and
        evidence_extract_llm, response_compose_llm, then grounding_guard_llm.
        All Agent LLM stages must go through the LLM Gateway for timeout,
        retry, JSON parse, schema validation, and fallback control.
        retrieve_sources reuses the existing search service when Zhihu config
        is available and writes a fallback stage/artifact otherwise.
        evidence_extract_llm produces an evidence artifact or rule fallback
        artifact without failing the task. response_compose_llm produces a
        final_result artifact or rule fallback artifact without failing the
        task. grounding_guard_llm produces a guarded_final_result artifact or
        rule fallback artifact without failing the task. Before completion,
        the worker writes a production_final_result artifact after deterministic
        sourceRefs validation. It does not run persona build, the demo search
        orchestrator, WebSocket/SSE, or stream events.
      operationId: createPersistentAgentTask
      x-implementation-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersistentAgentTaskCreateRequest"
            examples:
              default:
                value:
                  query: 不工作了之后，我想去新西兰生活
                  metadata: {}
      responses:
        "200":
          description: Persistent Agent task created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersistentAgentTaskCreateResponse"
        "400":
          description: Missing or invalid request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "503":
          description: DATABASE_URL or REDIS_URL is not configured, or Postgres/Redis is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
  /api/agent/tasks/{taskId}:
    get:
      tags:
        - Agent
      summary: Get current agent task state
      description: >
        First checks the legacy in-memory Agent task store used by
        /api/agent/search. If no in-memory task exists, checks the Phase 1
        persistent Postgres task store and returns product-facing task status,
        stage progress, partial/result availability, degraded flags, and
        structured error information.
      operationId: getDemoSearchAgentTask
      x-implementation-status: implemented
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Current in-memory Agent task or persistent Agent task status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentTaskResponse"
        "404":
          description: Task does not exist or expired from the in-memory store and was not found in Postgres.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
components:
  schemas:
...
    SearchSuccessResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: "#/components/schemas/SearchResult"
...
    ApiErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
    AgentTaskStartResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          required:
            - taskId
            - status
            - createdAt
          properties:
            taskId:
              type: string
            status:
              type: string
              const: running
            createdAt:
              type: string
              format: date-time
...
    PersistentAgentTaskListResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          required:
            - tasks
          properties:
            tasks:
              type: array
              items:
                $ref: "#/components/schemas/PersistentAgentTask"
...
    DemoSearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          minLength: 1
          description: Raw user question.
        count:
          type: integer
          minimum: 1
          maximum: 20
          default: 20
          description: Desired recall count before product-layer grouping.
        mode:
          type: string
          enum:
            - mock
            - cache_first
            - real
          default: mock
          description: Data mode. P0 must work in mock mode without Zhihu or LLM keys.
        sessionId:
          type: string
          description: Optional client-provided session id for local demo continuity.
...